change commandline options for file descriptors to established syntax for monitoring plugins add file descriptors to performance data
Jordans version of how to add thresholds for max file descriptors was nicer to read (on commandline and code-wise) but the new version sticks to the monitoring plugin guidelines and might be easier included into monitoring tools like Nagios and Icinga

While in the discussion Jordan thought wether there should be a check for open file descriptors or not it was already implemented so I adapted it to the common syntax for plugins. I'll add a issue to discuss wether disabling a check for open file descriptors should be implemented. Having this value in the performance data allows for adding a check like check_graphite which can check for spikes or unnatural rising in open file descriptors