change command line options to adopt monitoring plugin development guidelines raise version to 0.3.0
while it's a shame to remove @jordansissel 's code which was easier to use on the commandline and more sophisticated in the plugin the new version is better usable for monitoring admins since it uses the established syntax. This way it should also be more easily integrated into monitoring tools like Icinga or Nagios.

The current version is far from perfect but it is usable and already provides significant benefits over not monitoring Logstash via API.